update Dockerfile, rename files
