+++ src/predict/api/tests.py
@@ -1,1 +1,2 @@
+import pytest
